code/python/week-2/pyproject.toml
[project]
name = "week-2-workflows-agents"
version = "0.1.0"
description = "Module 2: Workflows & Agents"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rq>=1.15.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.0.0",
    "ruff>=0.1.0",
]
